Move resource tests to subtests.
