+++ internal/idp_test/util_test.go
@@ -1,4 +1,4 @@
-package idp
+package idp_test
 
 import (
 	"bytes"
